get module-contents to handle exported cards or fields internally
